// src/generated/graphql.tsx
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  bigint: any;
  json: any;
  timestamptz: any;
};
// ...
export type _Fuzzysearch_Directory = {
  __typename?: '_fuzzysearch_directory';
  company: Scalars['String'];
  department: Scalars['String'];
  id: Scalars['Int'];
  name: Scalars['String'];
};

/** aggregated selection of "_fuzzysearch.directory" */
export type _Fuzzysearch_Directory_Aggregate = {
  __typename?: '_fuzzysearch_directory_aggregate';
  aggregate?: Maybe<_Fuzzysearch_Directory_Aggregate_Fields>;
  nodes: Array<_Fuzzysearch_Directory>;
};

/** aggregate fields of "_fuzzysearch.directory" */
export type _Fuzzysearch_Directory_Aggregate_Fields = {
  __typename?: '_fuzzysearch_directory_aggregate_fields';
  avg?: Maybe<_Fuzzysearch_Directory_Avg_Fields>;
  count: Scalars['Int'];
  max?: Maybe<_Fuzzysearch_Directory_Max_Fields>;
  min?: Maybe<_Fuzzysearch_Directory_Min_Fields>;
  stddev?: Maybe<_Fuzzysearch_Directory_Stddev_Fields>;
  stddev_pop?: Maybe<_Fuzzysearch_Directory_Stddev_Pop_Fields>;
  stddev_samp?: Maybe<_Fuzzysearch_Directory_Stddev_Samp_Fields>;
  sum?: Maybe<_Fuzzysearch_Directory_Sum_Fields>;
  var_pop?: Maybe<_Fuzzysearch_Directory_Var_Pop_Fields>;
  var_samp?: Maybe<_Fuzzysearch_Directory_Var_Samp_Fields>;
  variance?: Maybe<_Fuzzysearch_Directory_Variance_Fields>;
};
// ...
export type Contact = {
  __typename?: 'contact';
  created_at: Scalars['timestamptz'];
  first_name: Scalars['String'];
  id: Scalars['Int'];
  last_name: Scalars['String'];
  /** An array relationship */
  phones: Array<Phone>;
  /** An aggregate relationship */
  phones_aggregate: Phone_Aggregate;
  updated_at?: Maybe<Scalars['timestamptz']>;
};
// ...
export type Contact_Aggregate = {
  __typename?: 'contact_aggregate';
  aggregate?: Maybe<Contact_Aggregate_Fields>;
  nodes: Array<Contact>;
};

/** aggregate fields of "contact" */
export type Contact_Aggregate_Fields = {
  __typename?: 'contact_aggregate_fields';
  avg?: Maybe<Contact_Avg_Fields>;
  count: Scalars['Int'];
  max?: Maybe<Contact_Max_Fields>;
  min?: Maybe<Contact_Min_Fields>;
  stddev?: Maybe<Contact_Stddev_Fields>;
  stddev_pop?: Maybe<Contact_Stddev_Pop_Fields>;
  stddev_samp?: Maybe<Contact_Stddev_Samp_Fields>;
  sum?: Maybe<Contact_Sum_Fields>;
  var_pop?: Maybe<Contact_Var_Pop_Fields>;
  var_samp?: Maybe<Contact_Var_Samp_Fields>;
  variance?: Maybe<Contact_Variance_Fields>;
};
// ...
export type Contact_Insert_Input = {
  created_at?: InputMaybe<Scalars['timestamptz']>;
  first_name?: InputMaybe<Scalars['String']>;
  id?: InputMaybe<Scalars['Int']>;
  last_name?: InputMaybe<Scalars['String']>;
  phones?: InputMaybe<Phone_Arr_Rel_Insert_Input>;
  updated_at?: InputMaybe<Scalars['timestamptz']>;
};
// ...
export type Contact_Obj_Rel_Insert_Input = {
  data: Contact_Insert_Input;
  /** upsert condition */
  on_conflict?: InputMaybe<Contact_On_Conflict>;
};

/** on_conflict condition type for table "contact" */
export type Contact_On_Conflict = {
  constraint: Contact_Constraint;
  update_columns?: Array<Contact_Update_Column>;
  where?: InputMaybe<Contact_Bool_Exp>;
};
// ...
export type Contact_Set_Input = {
  created_at?: InputMaybe<Scalars['timestamptz']>;
  first_name?: InputMaybe<Scalars['String']>;
  id?: InputMaybe<Scalars['Int']>;
  last_name?: InputMaybe<Scalars['String']>;
  updated_at?: InputMaybe<Scalars['timestamptz']>;
};
// ...
export type Elements = {
  __typename?: 'elements';
  element: Scalars['json'];
  id: Scalars['bigint'];
  /** An object relationship */
  project: Projects;
  project_id: Scalars['bigint'];
};


/** columns and relationships of "elements" */
export type ElementsElementArgs = {
  path?: InputMaybe<Scalars['String']>;
};

/** aggregated selection of "elements" */
export type Elements_Aggregate = {
  __typename?: 'elements_aggregate';
  aggregate?: Maybe<Elements_Aggregate_Fields>;
  nodes: Array<Elements>;
};

/** aggregate fields of "elements" */
export type Elements_Aggregate_Fields = {
  __typename?: 'elements_aggregate_fields';
  avg?: Maybe<Elements_Avg_Fields>;
  count: Scalars['Int'];
  max?: Maybe<Elements_Max_Fields>;
  min?: Maybe<Elements_Min_Fields>;
  stddev?: Maybe<Elements_Stddev_Fields>;
  stddev_pop?: Maybe<Elements_Stddev_Pop_Fields>;
  stddev_samp?: Maybe<Elements_Stddev_Samp_Fields>;
  sum?: Maybe<Elements_Sum_Fields>;
  var_pop?: Maybe<Elements_Var_Pop_Fields>;
  var_samp?: Maybe<Elements_Var_Samp_Fields>;
  variance?: Maybe<Elements_Variance_Fields>;
};
// ...
export type Phone = {
  __typename?: 'phone';
  /** An object relationship */
  contact?: Maybe<Contact>;
  contact_id: Scalars['Int'];
  created_ad?: Maybe<Scalars['timestamptz']>;
  id: Scalars['Int'];
  number: Scalars['String'];
};

/** aggregated selection of "phone" */
export type Phone_Aggregate = {
  __typename?: 'phone_aggregate';
  aggregate?: Maybe<Phone_Aggregate_Fields>;
  nodes: Array<Phone>;
};

/** aggregate fields of "phone" */
export type Phone_Aggregate_Fields = {
  __typename?: 'phone_aggregate_fields';
  avg?: Maybe<Phone_Avg_Fields>;
  count: Scalars['Int'];
  max?: Maybe<Phone_Max_Fields>;
  min?: Maybe<Phone_Min_Fields>;
  stddev?: Maybe<Phone_Stddev_Fields>;
  stddev_pop?: Maybe<Phone_Stddev_Pop_Fields>;
  stddev_samp?: Maybe<Phone_Stddev_Samp_Fields>;
  sum?: Maybe<Phone_Sum_Fields>;
  var_pop?: Maybe<Phone_Var_Pop_Fields>;
  var_samp?: Maybe<Phone_Var_Samp_Fields>;
  variance?: Maybe<Phone_Variance_Fields>;
};
// ...
export type Phone_Insert_Input = {
  contact?: InputMaybe<Contact_Obj_Rel_Insert_Input>;
  contact_id?: InputMaybe<Scalars['Int']>;
  created_ad?: InputMaybe<Scalars['timestamptz']>;
  id?: InputMaybe<Scalars['Int']>;
  number?: InputMaybe<Scalars['String']>;
};
// ...
export type Playground_Brands = {
  __typename?: 'playground_brands';
  brand_name: Scalars['String'];
  brand_owner: Scalars['String'];
  id: Scalars['Int'];
};

/** aggregated selection of "playground.brands" */
export type Playground_Brands_Aggregate = {
  __typename?: 'playground_brands_aggregate';
  aggregate?: Maybe<Playground_Brands_Aggregate_Fields>;
  nodes: Array<Playground_Brands>;
};

/** aggregate fields of "playground.brands" */
export type Playground_Brands_Aggregate_Fields = {
  __typename?: 'playground_brands_aggregate_fields';
  avg?: Maybe<Playground_Brands_Avg_Fields>;
  count: Scalars['Int'];
  max?: Maybe<Playground_Brands_Max_Fields>;
  min?: Maybe<Playground_Brands_Min_Fields>;
  stddev?: Maybe<Playground_Brands_Stddev_Fields>;
  stddev_pop?: Maybe<Playground_Brands_Stddev_Pop_Fields>;
  stddev_samp?: Maybe<Playground_Brands_Stddev_Samp_Fields>;
  sum?: Maybe<Playground_Brands_Sum_Fields>;
  var_pop?: Maybe<Playground_Brands_Var_Pop_Fields>;
  var_samp?: Maybe<Playground_Brands_Var_Samp_Fields>;
  variance?: Maybe<Playground_Brands_Variance_Fields>;
};
// ...
export type Projects = {
  __typename?: 'projects';
  id: Scalars['bigint'];
  name: Scalars['String'];
  slug: Scalars['String'];
  thumbnail?: Maybe<Scalars['String']>;
  updated_at: Scalars['timestamptz'];
};

/** aggregated selection of "projects" */
export type Projects_Aggregate = {
  __typename?: 'projects_aggregate';
  aggregate?: Maybe<Projects_Aggregate_Fields>;
  nodes: Array<Projects>;
};

/** aggregate fields of "projects" */
export type Projects_Aggregate_Fields = {
  __typename?: 'projects_aggregate_fields';
  avg?: Maybe<Projects_Avg_Fields>;
  count: Scalars['Int'];
  max?: Maybe<Projects_Max_Fields>;
  min?: Maybe<Projects_Min_Fields>;
  stddev?: Maybe<Projects_Stddev_Fields>;
  stddev_pop?: Maybe<Projects_Stddev_Pop_Fields>;
  stddev_samp?: Maybe<Projects_Stddev_Samp_Fields>;
  sum?: Maybe<Projects_Sum_Fields>;
  var_pop?: Maybe<Projects_Var_Pop_Fields>;
  var_samp?: Maybe<Projects_Var_Samp_Fields>;
  variance?: Maybe<Projects_Variance_Fields>;
};
// ...
export type Query_Root = {
  __typename?: 'query_root';
  /** fetch data from the table: "_fuzzysearch.directory" */
  _fuzzysearch_directory: Array<_Fuzzysearch_Directory>;
  /** fetch aggregated fields from the table: "_fuzzysearch.directory" */
  _fuzzysearch_directory_aggregate: _Fuzzysearch_Directory_Aggregate;
  /** fetch data from the table: "_fuzzysearch.directory" using primary key columns */
  _fuzzysearch_directory_by_pk?: Maybe<_Fuzzysearch_Directory>;
  /** execute function "_fuzzysearch.search" which returns "_fuzzysearch.directory" */
  _fuzzysearch_search: Array<_Fuzzysearch_Directory>;
  /** execute function "_fuzzysearch.search" and query aggregates on result of table type "_fuzzysearch.directory" */
  _fuzzysearch_search_aggregate: _Fuzzysearch_Directory_Aggregate;
  /** fetch data from the table: "contact" */
  contact: Array<Contact>;
  /** fetch aggregated fields from the table: "contact" */
  contact_aggregate: Contact_Aggregate;
  /** fetch data from the table: "contact" using primary key columns */
  contact_by_pk?: Maybe<Contact>;
  /** fetch data from the table: "elements" */
  elements: Array<Elements>;
  /** fetch aggregated fields from the table: "elements" */
  elements_aggregate: Elements_Aggregate;
  /** fetch data from the table: "elements" using primary key columns */
  elements_by_pk?: Maybe<Elements>;
  /** fetch data from the table: "phone" */
  phone: Array<Phone>;
  /** fetch aggregated fields from the table: "phone" */
  phone_aggregate: Phone_Aggregate;
  /** fetch data from the table: "phone" using primary key columns */
  phone_by_pk?: Maybe<Phone>;
  /** fetch data from the table: "playground.brands" */
  playground_brands: Array<Playground_Brands>;
  /** fetch aggregated fields from the table: "playground.brands" */
  playground_brands_aggregate: Playground_Brands_Aggregate;
  /** fetch data from the table: "playground.brands" using primary key columns */
  playground_brands_by_pk?: Maybe<Playground_Brands>;
  /** fetch data from the table: "projects" */
  projects: Array<Projects>;
  /** fetch aggregated fields from the table: "projects" */
  projects_aggregate: Projects_Aggregate;
  /** fetch data from the table: "projects" using primary key columns */
  projects_by_pk?: Maybe<Projects>;
};